+++ backend/users/Touir.ExpensesManager.Users/Controllers/EO/ApplicationEo.cs
@@ -8,5 +8,6 @@
         public string? Description { get; set; }
         public string? UrlPath { get; set; }
         public string? ResetPasswordUrlPath { get; set; }
+        public string? VerifyEmailErrorUrlPath { get; set; }
     }
 }